MB-26887: Move to bufferevent [1/7]
This patch is the first in a serie to move the connection
buffers over to use libevents bufferevents for the
network layer. It'll utilize the SSL support for bufferevents
so we'll also remove our own SSL layer on top of the sockets.

To reduce the size of the patch we've disabled some functionality:

* Disable fair sharing by backing off after processing n
  events

  This means that a single connection may starve the other
  connections unless it runs out of data in the input queue
  or the underlying engine needs to block.

  This will be enabled once the state machine is modified
  after the complete move to bufferevents

In addition to that no changes has been performed to the
input/output buffer used by memcached in order to keep
the patch as small as possible.

This means that libevent will read data into its buffers,
then memcached will copy the data into its own buffers
before using it (and the same happens when sending data).
This will be cleaned up in the next set of patches.

Change-Id: I4e7a7a95dc0535e4eb667819f31eb094bf86507b