Prepare various Command Context for OoO
When operating in OoO mode the underlying buffer for the
request may change the first time the command returns EWOULDBLOCK
as we'll try not to copy the buffer if we don't need our own
copy. This means that we can't keep pointers into the packet as
members in the command context (but fetch them from the cookie
every time).

Change-Id: Ib5516c5c162807e75140fc070c03e663e7e155c8